clean things up and fix csr issues
I made the chart check the devicePixelRatio so that it renders cleanly
on high density screens, but renders much fewer pixels on normal density
screens. This turned out to be more difficult than I thought, because we
need to access window which isn't available on the server. I had to add
a new hook, useCsr, to determine if we're on the server or client and
only run on the client.

* turns out there was a much easier way to do this, just update the width
and height from the renderChart function.